Add buildMetadata task and ref (#5511)
* Add buildMetadata task and ref

Move build metadata tasks, draft buildMetadata reference

Clean up buildMetadata ref

Add managed by label task

Add local non-generated task

Add local generated resource

Add remote generator task

Clean up tasks and ref

Add local transformer annotation example

Add local and remote transformer example

* Address PR feedback and general cleanup

cherrypick updates from feature branch

fix script

fix release script

+++ releasing/helpers.sh
@@ -2,12 +2,20 @@
 # Copyright 2022 The Kubernetes Authors.
 # SPDX-License-Identifier: Apache-2.0
 
+ORIGIN_MASTER="feat/5449-add-release-automation"
+UPSTREAM_REPO="upstream"
+UPSTREAM_MASTER="master"
 
 function createBranch {
   branch=$1
   title=$2
   echo "Making branch $branch : \"$title\""
-  git branch -D $branch  # delete if it exists
+  # Check if release branch exists
+  if git show-ref --quiet "refs/heads/${branch}"; then
+    git fetch --tags upstream master
+    git checkout $ORIGIN_MASTER
+    git branch -D $branch  # delete if it exists
+  fi
   git checkout -b $branch
   git commit -a -m "$title"
   git push -f origin $branch
